Update 2
Add event listener to the buttons, and see which one was clicked, changing the text and colour of the button depending on the case. Also add check to see if number entered was at least the minimum amount, with the effect of changing the submit button text and colour depending on the case. Once a valid submit is entered, the input filed is disabled and the button changes text and colour again. If clicked again the page refreshes and the game can be played again.